add: maven, gson
This is a rather huge commit, with refactorings all over the place. The very basic Maven Architecture was set up. `org.json` has been replaced with `com.google.code.gson` for reasons, such as better functionality and higher maintenance!
A changelog was started to keep track of this PR's changes, since there likely will be a lot.
Application now compiles and runs with a few minor appearance issues, which will be fixed in the coming commits!